+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Attendance">
         <activity
-            android:name="com.yapp.presentation.ui.login.LoginActivity"
+            android:name="com.yapp.presentation.ui.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.Attendance.NoActionBar">
             <intent-filter>
@@ -21,12 +21,6 @@
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-
-        <activity
-            android:name="com.yapp.presentation.ui.member.main.QRMainActivity"
-            android:exported="true"
-            android:theme="@style/Theme.Attendance.NoActionBar">
         </activity>
 
         <activity
