+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">attendance-android</string>
+    <string name="kakao_app_key">72806b8322ff87c0ab3ec377e4f99e6a</string>
 </resources>
