--- common/src/main/java/com/yapp/common/base/BaseScreen.kt
@@ -1,7 +1,0 @@
-package com.yapp.common.base
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun BaseScreen() {
-}
