+++ presentation/src/main/java/com/yapp/presentation/ui/MainActivity.kt
@@ -1,0 +1,19 @@
+package com.yapp.presentation.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import com.yapp.common.theme.AttendanceTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            AttendanceTheme {
+                AttendanceScreen()
+            }
+        }
+    }
+}
