+++ presentation/src/main/java/com/yapp/presentation/ui/member/main/QRMain.kt
@@ -1,17 +1,23 @@
-package com.yapp.presentation.ui.member.main.screen
+package com.yapp.presentation.ui.member.main
 
+import android.widget.Toast
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.material.*
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.hilt.navigation.compose.hiltViewModel
+import com.yapp.presentation.ui.login.state.LoginContract
 import com.yapp.presentation.ui.member.main.QRMainViewModel
+import com.yapp.presentation.ui.member.main.state.QRMainContract
 import com.yapp.presentation.ui.member.main.state.QRMainContract.*
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 
 @Composable
@@ -20,6 +26,7 @@ fun Main(
     coroutineScope: CoroutineScope = rememberCoroutineScope(),
     viewModel: QRMainViewModel = hiltViewModel(),
 ) {
+    val context = LocalContext.current
     val uiState = viewModel.uiState.collectAsState()
 
     Scaffold(
@@ -27,6 +34,16 @@ fun Main(
         modifier = Modifier
             .fillMaxSize()
     ) {
+        LaunchedEffect(key1 = viewModel.effect) {
+            viewModel.effect.collect { effect ->
+                when (effect) {
+                    is QRMainUiSideEffect.ShowToast -> {
+                        Toast.makeText(context, effect.msg, Toast.LENGTH_SHORT).show()
+                    }
+                }
+            }
+        }
+
         Column {
             Text(text = uiState.value.time)
